Create prime,composite pat
Write a C program using while loop to read the numbers until -1 is encountered. Count the number of prime and composite numbers entered by the user. 

Input: Numbers are given one after the other in separate lines. Give the last number as -1. 

Output: Print the count of the prime numbers using the variable name PrimeCount and the count of compostie numbers using the variable name CompositeCount.