Create bankaccount cat1 theory
suppose you are working on a program that models a banking system. you are asked to create a class called bank account that represents a single bank account, with following attributes:
1 account number
2 account holder name: as a string
3 balance: a double that represents the current balance in the account
The bank account class should have the following public member functions:
1 deposit: an inline function that takes an amount in double datatype and adds it to balance in the account.
2 withdraw: an inline function that takesan amount in double dattype and subtractsit from the balance in the account,if the account has sufficient funds. if the account does not have sufficient funds, it should print an error message and not update the balance.
3 get balance: an inline function that returns the current balance in the account.
Write a c++ program in which in the main function, create twobank accountobjects with different account numbers and account holder names, and initialize them with initail balances. prompt the user to enter a deposit and a withdrawl amount for each account,and use the deposit an dwithdraw functions to update the balances accordingly. Finally print out the current balance for both accounts.