Create magic bag
A magic person wants to put some data of the same type into his magic bag. He wants to repeat the
same magic with different data such as int, string, float etc. could be put into his bag.
The process of putting the data inside the bag is said to be as “Push”. The bag contains some restriction in its size. Ie. The size must be specified in the initial part of every iteration.Once when the input crosses the maximum limit it displays the message "Bag is full"
Implement this magic person’s game for his specification of any datatype being used inside his bag for each iteration.
Note that he should not give clue of doing this operation to anyone. So have the same name for the function that does the push operation for every datatype. Write a code in c++ sample input
5
10
20
30
40
50
60
Output 
Bag is full