Create cat 1 lab
Ms. Ananthidas generated the mark list for the class X-A1 in which she had mapped the name, roll number, date of birth of the students to the marks. 
Similarly she had generated the class X-B1 marksheet too. Get those Physics marks from the Ms. Ananthidas as input. 
i) Use appropriate datastructure in C Programming to store the values. (Use functions concept)
ii) Now generate a list of marks alone from that given values for X-A1 and for X-B1 separately and display the list (Pass the required data to the function and perform the operation). 
iii) She now wants to combine the rank lists together in the descending order and have the consolidated list of marks of her students. Do not return the revised descending ordered data from the function. Use appropriate concept for that.
iv) She wants to know all the details of the students who have got first mark and last mark after combining the marks of both the classes.
Note: If two students have got the same mark, then do the following to arrange in descending order of the marks
- Check the date of birth of both the students and the one with higher age among the two will be given the top rank.
Help her to do the same using your C program.